fix: refatora componentes
