+++ src/components/removeMonitorModal/index.tsx
@@ -44,7 +44,9 @@ const RemoveMonitorModal = ({
 
           <ConfirmationTextContainer>
             <Typography variant="h4">Tudo certo!</Typography>
-            <Typography variant="body1">O(A) monitor(a) foi removido(a)</Typography>
+            <Typography variant="body1">
+              O(A) monitor(a) foi removido(a)
+            </Typography>
           </ConfirmationTextContainer>
 
           <Button onClick={handleClose} color="primary">
